Move distinct and raw symmetry to cana/symmetry.py Move symmetry tests to tests/test_symmetry.py
