+++ cana/boolean_node.py
@@ -14,7 +14,7 @@ Main class for Boolean node objects.
 #   MIT license.
 from __future__ import division
 
-from itertools import combinations, compress, permutations, product
+from itertools import combinations, compress, product
 from statistics import mean
 
 import networkx as nx
@@ -23,6 +23,7 @@ import pandas as pd
 
 import cana.canalization.boolean_canalization as BCanalization
 import cana.canalization.cboolean_canalization as cBCanalization
+import cana.symmetry as symmetry
 from cana.cutils import (
     binstate_to_statenum,
     flip_bit,
@@ -407,41 +408,11 @@ class BooleanNode(object):
 
         Returns:
             (float)
+
+        See also:
+            :func:`cana.symmetry.distinct_symmetry`
         """
-        if not self.outputs:
-            return 0.0
-
-        lut = list(map(str, self.outputs))
-        total_ratio = 0.0
-        row_count = len(lut)
-
-        # Rows with the same number of 1s share the same distinct permutations.
-        perm_cache = {}
-
-        for index, output_symbol in enumerate(lut):
-            input_bits = statenum_to_binstate(index, base=self.k)
-            cache_key = (len(input_bits), input_bits.count("1"))
-
-            if cache_key not in perm_cache:
-                perm_cache[cache_key] = tuple(
-                    sorted({"".join(perm) for perm in permutations(input_bits)})
-                )
-
-            distinct_perms = perm_cache[cache_key]
-            total_perms = len(distinct_perms)
-
-            if total_perms <= 1:
-                continue
-
-            matches = 0
-            for perm_bits in distinct_perms:
-                perm_index = int(perm_bits, 2)
-                if lut[perm_index] == output_symbol:
-                    matches += 1
-
-            total_ratio += (matches - 1) / (total_perms - 1)
-
-        return total_ratio / row_count if row_count else 0.0
+        return symmetry.distinct_symmetry(self.outputs, self.k)
 
     def raw_symmetry(self):
         """Compute the raw symmetry of the node LUT.
@@ -452,35 +423,11 @@ class BooleanNode(object):
 
         Returns:
             (float)
+
+        See also:
+            :func:`cana.symmetry.raw_symmetry`
         """
-        if not self.outputs:
-            return 0.0
-
-        lut = list(map(str, self.outputs))
-        rows_by_weight = {}
-
-        for index, output_symbol in enumerate(lut):
-            input_bits = statenum_to_binstate(index, base=self.k)
-            weight = input_bits.count("1")
-            rows_by_weight.setdefault(weight, []).append(output_symbol)
-
-        total_ratio = 0.0
-        total_rows = 0
-
-        for symbols in rows_by_weight.values():
-            group_size = len(symbols)
-            if group_size == 0:
-                continue
-
-            counts = {}
-            for symbol in symbols:
-                counts[symbol] = counts.get(symbol, 0) + 1
-
-            for symbol in symbols:
-                total_ratio += counts[symbol] / group_size
-                total_rows += 1
-
-        return total_ratio / total_rows if total_rows else 0.0
+        return symmetry.raw_symmetry(self.outputs, self.k)
 
     def look_up_table(self):
         """Returns the Look Up Table (LUT)
